Change nested component structure and update componnents state
