+++ src/App.js
@@ -11,6 +11,9 @@ import Hero from './components/Hero';
 import Footer from './components/Footer';
 import ProductDetail from './components/ProductDetail';
 import Checkout from './components/Checkout';
+import ProductList from './components/ProductList';
+import Analytics from './components/Analytics';
+import CartPopup from './components/cartPopup';
 
 function App() {
   const dispatch = useDispatch();
@@ -18,6 +21,16 @@ function App() {
   const carts = useSelector((state) => state.carts);
 
   const [counter, setCounter] = useState(0);
+
+  const [isOpen, setIsOpen] = React.useState(false);
+
+  const handleOpen = () => {
+    setIsOpen(true);
+  }
+
+  const isClose = () => {
+    setIsOpen(false);
+  }
 
 
   const increment = () => {
@@ -28,6 +41,20 @@ function App() {
     setCounter(counter - 1);
   };
 
+  const [total, setTotal] = useState(0);
+
+  const cartItems = useSelector((state) => state.carts);
+  const cartProducts = cartItems;
+
+ useEffect(() => {
+    const total = cartProducts.reduce((acc, product) => {
+      const price = product.price;
+      const quantity = product.quantity;
+      return acc + (price * quantity);
+    }, 0);
+    setTotal(total);
+  }, [total, counter]);
+
   useEffect(() => {
     dispatch(getProducts());
     dispatch(fetchCart());
@@ -36,13 +63,26 @@ function App() {
   
   return (
     <div>
-      <Navbar counter={counter} decrement={decrement} />
-        <Routes>
-          <Route path="/" element={<Hero increment={increment} />} />
-          {/* <Route path="/" element={<ProductList />} /> */}
-          <Route path="/products/:id" element={<ProductDetail />} />
-          <Route path="/checkout" element={<Checkout />} />
-        </Routes>
+      <Navbar counter={counter} handleOpen={handleOpen} />
+      <Routes>
+        <Route path="/" element={
+          <>
+          <Hero />
+          <ProductList increment={increment} />
+          <Analytics />
+          </>
+          } 
+        />
+        <Route path="/products/:id" element={<ProductDetail />} />
+        <Route path="/checkout" element={<Checkout cartItems={cartItems} total={total} />} />
+      </Routes>
+      {isOpen && <CartPopup 
+        isOpen={isOpen}
+        isClose={isClose}
+        decrement={decrement}
+        cartItems={cartItems}
+        total={total}
+      /> }
       <Footer />
     </div>
   );
